varios servicios terminados - explicado en daily

+++ DadoVentas/src/main/java/com/DadoSoft/DadoVentas/repositorio/ProductosRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductosRepository extends JpaRepository<Productos, Integer> {
+public interface ProductosRepository extends JpaRepository<Productos, Long> {
 
     @Query(nativeQuery = true, value = "SELECT * FROM Productos order by desc limit 1")
     public Productos ObtenerUltimoProducto();
